# Improve psql performance by isolating uniqueness query.