@stephenafamo 2022-06-19
This is very nice. Thank you! All the tests are passing, so I'll go ahead and merge this.

Once #1143 is also completed and merged, then there would be a significant speed boost overall.